+++ syft/__init__.py
@@ -37,8 +37,8 @@ from syft.frameworks.torch import TorchHook
 # Import Tensor Types
 from syft.frameworks.torch.tensors.decorators import LoggingTensor
 from syft.frameworks.torch.tensors.interpreters import AdditiveSharingTensor
-from syft.frameworks.torch.tensors.interpreters import MultiPointerTensor
 from syft.frameworks.torch.tensors.interpreters import AutogradTensor
+from syft.generic.pointers import MultiPointerTensor
 from syft.generic.pointers import PointerTensor
 
 # import other useful classes
@@ -71,7 +71,6 @@ from syft.frameworks.torch.tensors.interpreters import CRTPrecisionTensor
 from syft.frameworks.torch.tensors.interpreters import AutogradTensor
 from syft.frameworks.torch.tensors.interpreters import FixedPrecisionTensor
 from syft.frameworks.torch.tensors.interpreters import LargePrecisionTensor
-from syft.frameworks.torch.tensors.interpreters import MultiPointerTensor
 
 from syft.generic.pointers import ObjectPointer
 from syft.generic.pointers import CallablePointer
