+++ syft/frameworks/torch/hook/hook_args.py
@@ -13,11 +13,11 @@ from syft.frameworks.torch.tensors.interpreters import AutogradTensor
 from syft.frameworks.torch.tensors.interpreters import TorchTensor
 from syft.frameworks.torch.tensors.interpreters import FixedPrecisionTensor
 from syft.frameworks.torch.tensors.interpreters import AdditiveSharingTensor
-from syft.frameworks.torch.tensors.interpreters import MultiPointerTensor
 from syft.frameworks.torch.tensors.interpreters import CRTPrecisionTensor
 from syft.frameworks.torch.tensors.interpreters import LargePrecisionTensor
 from syft.frameworks.torch.tensors.decorators import LoggingTensor
 from syft.generic.tensor import AbstractTensor
+from syft.generic.pointers import MultiPointerTensor
 from syft.generic.pointers import PointerTensor
 from syft.workers import AbstractWorker
