+++ syft/frameworks/torch/tensors/interpreters/__init__.py
@@ -1,7 +1,6 @@
 from .native import TorchTensor  # noqa: F401
 from .precision import FixedPrecisionTensor  # noqa: F401
 from .additive_shared import AdditiveSharingTensor  # noqa: F401
-from .multi_pointer import MultiPointerTensor  # noqa: F401
 from .crt_precision import CRTPrecisionTensor  # noqa: F401
 from .large_precision import LargePrecisionTensor  # noqa: F401
 from .autograd import AutogradTensor  # noqa: F401
