+++ syft/generic/pointers/__init__.py
@@ -1,7 +1,12 @@
 from syft.generic.pointers.object_pointer import ObjectPointer
 from syft.generic.pointers.callable_pointer import create_callable_pointer
 from syft.generic.pointers.callable_pointer import CallablePointer
+from syft.generic.pointers.multi_pointer import MultiPointerTensor
 from syft.generic.pointers.pointer_tensor import PointerTensor
 from syft.generic.pointers.object_wrapper import ObjectWrapper
 
-__all__ = ["ObjectPointer", "CallablePointer", "create_callable_pointer", "PointerTensor"]
+__all__ = ["ObjectPointer",
+           "CallablePointer",
+           "create_callable_pointer",
+           "MultiPointerTensor",
+           "PointerTensor"]
