+++ syft/generic/pointers/multi_pointer.py
